pyproject.toml
[tool.poetry]
name = "web-scraper"
version = "0.1.0"
description = ""
authors = ["Edward Regalado <tregalado30@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
